Move util into forecast.
